sort out some classes and fix some warnings
sort out some classes and fix some warnings
