+++ eccoin-qt.pro
@@ -286,7 +286,8 @@ HEADERS += \
     src/script/stakescript.h \
     src/messages.h \
     src/processblock.h \
-    src/processheader.h
+    src/processheader.h \
+    src/blockindex.h
 
 
 # organize compiles of cpp files by section, this seems to be a logical order where the files lower down generally depend
@@ -387,7 +388,8 @@ SOURCES += \
     src/rpcwallet.cpp \
     src/messages.cpp \
     src/processblock.cpp \
-    src/processheader.cpp
+    src/processheader.cpp \
+    src/blockindex.cpp
