+++ src/chain.cpp
@@ -58,93 +58,3 @@ const CBlockIndex *CChain::FindFork(const CBlockIndex *pindex) const {
         pindex = pindex->pprev;
     return pindex;
 }
-
-/** Turn the lowest '1' bit in the binary representation of a number into a '0'. */
-int static inline InvertLowestOne(int n) { return n & (n - 1); }
-
-/** Compute what height to jump back to with the CBlockIndex::pskip pointer. */
-int static inline GetSkipHeight(int height) {
-    if (height < 2)
-        return 0;
-
-    // Determine which height to jump back to. Any number strictly lower than height is acceptable,
-    // but the following expression seems to perform well in simulations (max 110 steps to go back
-    // up to 2**18 blocks).
-    return (height & 1) ? InvertLowestOne(InvertLowestOne(height - 1)) + 1 : InvertLowestOne(height);
-}
-
-CBlockIndex* CBlockIndex::GetAncestor(int height)
-{
-    if (height > nHeight || height < 0)
-        return NULL;
-
-    CBlockIndex* pindexWalk = this;
-    int heightWalk = nHeight;
-    while (heightWalk > height) {
-        int heightSkip = GetSkipHeight(heightWalk);
-        int heightSkipPrev = GetSkipHeight(heightWalk - 1);
-        if (pindexWalk->pskip != NULL &&
-            (heightSkip == height ||
-             (heightSkip > height && !(heightSkipPrev < heightSkip - 2 &&
-                                       heightSkipPrev >= height)))) {
-            // Only follow pskip if pprev->pskip isn't better than pskip->pprev.
-            pindexWalk = pindexWalk->pskip;
-            heightWalk = heightSkip;
-        } else {
-            pindexWalk = pindexWalk->pprev;
-            heightWalk--;
-        }
-    }
-    return pindexWalk;
-}
-
-const CBlockIndex* CBlockIndex::GetAncestor(int height) const
-{
-    return const_cast<CBlockIndex*>(this)->GetAncestor(height);
-}
-
-void CBlockIndex::BuildSkip()
-{
-    if (pprev)
-        pskip = pprev->GetAncestor(GetSkipHeight(nHeight));
-}
-
-bool CBlockIndex::IsProofOfWork() const
-{
-    return !(nFlags & BLOCK_PROOF_OF_STAKE);
-}
-
-bool CBlockIndex::IsProofOfStake() const
-{
-    return (nFlags & BLOCK_PROOF_OF_STAKE);
-}
-
-void CBlockIndex::SetProofOfStake()
-{
-    nFlags |= BLOCK_PROOF_OF_STAKE;
-}
-
-unsigned int CBlockIndex::GetStakeEntropyBit() const
-{
-    return ((nFlags & BLOCK_STAKE_ENTROPY) >> 1);
-}
-
-bool CBlockIndex::SetStakeEntropyBit(unsigned int nEntropyBit)
-{
-    if (nEntropyBit > 1)
-        return false;
-    nFlags |= (nEntropyBit? BLOCK_STAKE_ENTROPY : 0);
-    return true;
-}
-
-bool CBlockIndex::GeneratedStakeModifier() const
-{
-    return (nFlags & BLOCK_STAKE_MODIFIER);
-}
-
-void CBlockIndex::SetStakeModifier(uint64_t nModifier, bool fGeneratedStakeModifier)
-{
-    nStakeModifier = nModifier;
-    if (fGeneratedStakeModifier)
-        nFlags |= BLOCK_STAKE_MODIFIER;
-}
