+++ src/checkpoints.cpp
@@ -29,11 +29,11 @@ namespace Checkpoints {
         if (pindex==NULL)
             return 0.0;
 
-        int64_t nNow = time(NULL);
+//        int64_t nNow = time(NULL);
 
-        double fSigcheckVerificationFactor = fSigchecks ? SIGCHECK_VERIFICATION_FACTOR : 1.0;
-        double fWorkBefore = 0.0; // Amount of work done before pindex
-        double fWorkAfter = 0.0;  // Amount of work left after pindex (estimated)
+//        double fSigcheckVerificationFactor = fSigchecks ? SIGCHECK_VERIFICATION_FACTOR : 1.0;
+//        double fWorkBefore = 0.0; // Amount of work done before pindex
+//        double fWorkAfter = 0.0;  // Amount of work left after pindex (estimated)
         // Work is defined as: 1.0 per transaction before the last checkpoint, and
         // fSigcheckVerificationFactor per transaction after.
 /*
